IDs rewritten to Longs

+++ src/main/java/com/elte/supplymanagersystem/controllers/CompanyController.java
@@ -52,7 +52,7 @@ public class CompanyController {
      * @return Returns a ResponseEntity of the Company with the given ID.
      */
     @GetMapping("/{id}")
-    public ResponseEntity get(@PathVariable Integer id, Authentication auth) {
+    public ResponseEntity get(@PathVariable Long id, Authentication auth) {
         User loggedInUser = userService.getValidUser(auth.getName());
         if (loggedInUser != null) {
             return companyService.getById(loggedInUser, id);
@@ -87,7 +87,7 @@ public class CompanyController {
      */
     //Update
     @PutMapping("/{id}")
-    public ResponseEntity put(@RequestBody CompanyDTO companyDTO, @PathVariable Integer id, Authentication auth) {
+    public ResponseEntity put(@RequestBody CompanyDTO companyDTO, @PathVariable Long id, Authentication auth) {
         User loggedInUser = userService.getValidUser(auth.getName());
         if (loggedInUser != null) {
             return companyService.putById(companyDTO, loggedInUser, id);
@@ -122,7 +122,7 @@ public class CompanyController {
      * @return Returns a ResponseEntity: OK if the deletion was successful and NotFound if the record was not found.
      */
     @PutMapping("/{id}/disable")
-    public ResponseEntity disable(@PathVariable Integer id, Authentication auth) {
+    public ResponseEntity disable(@PathVariable Long id, Authentication auth) {
         User loggedInUser = userService.getValidUser(auth.getName());
         if (loggedInUser != null) {
             return companyService.disableCompany(id, loggedInUser);
@@ -139,7 +139,7 @@ public class CompanyController {
      * @return Returns a ResponseEntity: OK if the deletion was successful and NotFound if the record was not found.
      */
     @PutMapping("/{id}/enable")
-    public ResponseEntity enable(@PathVariable Integer id, Authentication auth) {
+    public ResponseEntity enable(@PathVariable Long id, Authentication auth) {
         User loggedInUser = userService.getValidUser(auth.getName());
         if (loggedInUser != null) {
             return companyService.enableCompany(id, loggedInUser);
@@ -157,7 +157,7 @@ public class CompanyController {
      */
     //Delete
     @DeleteMapping("/{id}")
-    public ResponseEntity delete(@PathVariable Integer id, Authentication auth) {
+    public ResponseEntity delete(@PathVariable Long id, Authentication auth) {
         User loggedInUser = userService.getValidUser(auth.getName());
         if (loggedInUser != null) {
             return companyService.deleteById(id, loggedInUser);
